refactor: move Hand construction internals out of public header
Reviewer (zengguizhang) pointed out that the WUJIHANDCPP_API tags we
added to probe_handedness, the SN registry helpers, and the Handler
SN accessor were band-aids for a deeper layering issue: Hand's
non-trivial ctor/dtor lived inline in the public header, which forced
private helpers to leak into the ABI surface. Address the root cause:

1. Hand(serial_number=...), Hand(Side, ...), and ~Hand() bodies move
   to src/device/hand.cpp. The header now has just three exported
   declarations (still WUJIHANDCPP_API for shared-library users) and
   keeps the rest of the inline template-heavy interface untouched.

2. The registry helpers (register_hand_sn / unregister_hand_sn /
   held_sns_snapshot) drop out of the public detail:: namespace
   entirely. They live in hand.cpp's anonymous namespace, callable
   only from within the library.

3. probe_handedness loses its Hand::private-static-with-API marker
   and becomes a free function in hand.cpp's anonymous namespace,
   same access scope as the registry.

4. Hand::SnRegistration keeps its struct definition (it's still a
   member of Hand) but the destructor is now out-of-line in
   hand.cpp, so the unregister_hand_sn call doesn't need to be
   visible at the call site in the header.

5. Handler::selected_serial_number() moves to Handler's private
   section + becomes a friend-only accessor: protocol::Handler adds
   "friend class wujihandcpp::device::Hand;" and the WUJIHANDCPP_API
   marker drops. Hand is the sole intended caller (for SN-registry
   bookkeeping), and external C++ consumers can't call it because
   the symbol is hidden from the shared library.

   This required forward-declaring wujihandcpp::device::Hand in
   handler.hpp; it remains an internal-only access channel.

The retired dedicated test file wujihandcpp/tests/device/
hand_registry_test.cpp is removed: those cases tested 3 tiny
unordered_set wrapper functions that are no longer reachable from
test scope, and the underlying registry behavior is implicitly
covered by probe_handedness (logic) + HIL (multi-hand scenario).

Public-header surface diff: -132 / +21. Public ABI surface dropped
4 exported symbols (probe_handedness, 3 registry helpers, Handler
selected_serial_number) while keeping all user-facing functionality
intact. The full set of changes is verified by:
  - C++ 26/26 + Python 12/12 + C++11 header-compat build
  - HIL: single-hand correct/wrong side, double-hand pair,
    timing-double-device all pass

Resolve m-6993430758
